feat!: new disparity input format
In the input section of configuration files, keys `disp_min` and
`disp_max` are merged into a single key `disp_left` while keys
`disp_min_right` and `disp_max_right` are merged into a `disp_right`
key.

Values should be a tuple of integers with min and max value in this
order or a single string representing path to a disparity grid file.
This file must be an image in tif format with two bands: first band
corresponds to the min grid and is labeled `min` while second band
corresponds to the max grid and is labeled `max`.

Such files are added to test data as:

  - `tests/pandora/left_disparity_grid.tif
  - `tests/pandora/right_disparity_grid.tif`
  - `tests/pandora/tiny_left_disparity_grid.tif`

`disp_min` and `disp_max` attributes of disparity map datasets are
merged into a `disparity_interval` data_var with coordinates `min` and
`max`.

BREAKING CHANGE: inputs section of configuration files is changed.

Refs: #330