tests: move cv_masked tests to test_maching_cost
Tests of cv_masked from:

- tests/test_matching_cost_census.py
- tests/test_matching_cost_sad.py
- tests/test_matching_cost_zncc.py

are moved to tests/test_matching_cost.py.

Those tests are parametrized and executed on the four matching cost
methodes except for `census` when test involves a window size of 1.

Tests do not check the whole cost volume anymore but focus on the
expected positions of NaNs in cost volume after mask application. This
method is more in line with the objective of tests while it permits to
use same data and expected result with all cost volume methods.

Refs: #372