refact: move to src layout
Because the design of the tests requires them to import `tests.common.py`,
the tests folder must be a package, necessitating an `__init__.py` module
at its root. As a result of this configuration, pytest, by default,
makes the source package importable, which masks the package installed
in the virtual environment. This behavior prevents testing wheels.

The [documentation of pytest](https://docs.pytest.org/en/6.2.x/goodpractices.html#src-layout)
therefore recommends to use the `src-layout`.