chore: declare that pandora is typed
Due to the change of layout, mypy considers pandora to be
an external library so I added the `py.typed` [marker](https://mypy.readthedocs.io/en/stable/running_mypy.html#missing-imports).
I had to change `meson.build` in order the file `py.typed` to be
included into the wheel so that when used as a library, pandora will be
declared typed.

Unfortunately, mypy and pylint still does not found pandora when
checking tests. This bug is referenced for [setuptools](https://github.com/python/mypy/issues/13392), it could be the
same for meson-python but tagged as `wontfix`:

https://github.com/mesonbuild/meson-python/issues/555